ajout memset

+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: suvitiel <suvitiel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/08 22:24:57 by suvitiel          #+#    #+#             */
-/*   Updated: 2016/11/27 21:24:55 by suvitiel         ###   ########.fr       */
+/*   Updated: 2016/11/29 00:24:12 by suvitiel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 void	*ft_memset(void *s, int c, size_t n)
 {
-	unsigned int	count;
-	char			*str;
+	size_t	count;
+	char	*str;
 
 	count = 0;
 	str = s;
